[!][FEATURE] Add multiline and hovered map labels
New functionality:
* This adds multiline labels (one primary line and multiple secondary
  lines) to map objects
* All labels can have 2 states: hovered and normal
* Allows to configure label contents for airports, pilots and sectors
  (see preferences dialog) (more possible on request)
* Introduces a placeholder system for content in labels (E.g.
  `{atisLetter}`).
  Labels include {livestream} with hints to streams on twitch, youtube
  or owncast.
* Hovered labels have received a background rectangle to make them
  stand out more. The color of that backdrop is configurable and
  depending on if the label is of light or dark color.
* This also adds preferences for several old things that were missing
  options.
* QuteScoop now reads ratings (controller/pilot/military) from the
  VATSIM API.
* You can use QuteScoop to resolve flight plans to route waypoints from
  the console: ‹QuteScoop --route WSSS SURGA M635 UDONO WADD› prints the
  waypoints with some information. While originally intended for testing
  maybe this is useful for you.
* Friends are now shown in 2 new ways, too (the friends highlighter is
  still there):
  * Airport and pilot dots are drawn in the friends highlighter
    color (or you can set friends/pilotDotColor friends/airportDotColor)
    and slightly bigger.
  * Labels of friends (or labels of airports with befriended
    controllers...) get a rectangle in the highlighter color (or as
    overridden with friends/pilotLabelRectColor,
    friends/airportLabelRectColor and friends/sectorLabelRectColor).

Caching of labels has been improved which greatly improves rendering
performance. But on the other hand I used a function to measure drawn
texts (QFontMetrics::tightBoundingRect()) that is known to be slow on
Windows... If it became annoyingly slow for you please shoot me a message
and we'll work it out.

A general advise: Our label system is meant to be used with a hard limit
for number of labels (under Display → OpenGL graphics settings)
(Default: 130).
Labels are rendered by priority groups (basically sectors → airports
(more congested first) → pilots (friends first) → inactive airports →
waypoints) but only while the zoom level is above a set threshold for
that group.
Labels are still shown when you hover over any object.
As an example: That system shows more pilot labels in less dense areas
but does not show any when you are looking at Central Europe in evening
traffic.
The new function "View → Only show short-term route parts" also helps
with clutter.

Tooltips on the map have been removed by default because they were
interfering with the new UI IMHO. If you want to bring them back, set
mapUi/showTooltips=true in the config.

Default styles have been adapted, because everything looks better on a
darker Earth...

This also enables multisample anti-aliasing. Disable in the config
with gl/samplebuffers=false, set number of samples with gl/samples
(default: 4xMSAA).

A lot more changes, fixes and refactoring have gone into this commit...
Sorry that it became so convoluted.

Breaking changes:
* The workaround "Simple labels" (quick rendering in 3D without overlap
  checking) has been removed. I believe it is not needed with current
  hardware. I hope you don't mind!
* The ability to let QuteScoop take screenshots has been removed.
* The ability to reset parts of the configuration to default has been
  removed from the preferences dialog. This has always been inconsistent
  and proved to be a burden when moving things around.
  (Manual way: Save your settings to .ini, choose "reset all on next
  start", then import your edited .ini)

Some more backstory for developers: I tried to allow more flexibility for
the label system by using a JavaScript engine instead of C++ lambda
functions but it was basically just too slow for our needs.
And I finally fixed our dependency chain and I removed the inheritance
Client → MapObject. A rebuild of our search system(s) which is utter chaos
is planned.

Fixes: #242
Fixes: #252
Fixes: #242
Fixes: #263
Fixes: #49
Fixes: #264
Relates: #243

+++ src/NavData.h
@@ -15,27 +15,36 @@ enum LatLngPrecission {
     Secs = 3, Mins = 2, Degrees = 1
 };
 
-class NavData: public QObject {
+class NavData
+    : public QObject {
     Q_OBJECT
     public:
         static NavData* instance(bool createIfNoInstance = true);
         static QPair<double, double>* fromArinc(const QString &str);
-        static QString toArinc(const short lat, const short lon);
+        static QString toArinc(const float lat, const float lon);
         static QString toEurocontrol(const double lat, const double lon, const LatLngPrecission maxPrecision = LatLngPrecission::Secs);
 
         static double distance(double lat1, double lon1, double lat2, double lon2);
         static QPair<double, double> pointDistanceBearing(
-        double lat, double lon,
-        double dist, double heading
+            double lat,
+            double lon,
+            double dist,
+            double heading
         );
         static double courseTo(double lat1, double lon1, double lat2, double lon2);
         static QPair<double, double> greatCircleFraction(
-        double lat1, double lon1,
-        double lat2, double lon2, double fraction
+            double lat1,
+            double lon1,
+            double lat2,
+            double lon2,
+            double fraction
         );
         static QList<QPair<double, double> > greatCirclePoints(
-        double lat1, double lon1, double lat2,
-        double lon2, double intervalNm = 30.
+            double lat1,
+            double lon1,
+            double lat2,
+            double lon2,
+            double intervalNm = 30.
         );
         static void plotGreatCirclePoints(const QList<QPair<double, double> > &points);
